error-based-sqli/requirements.txt
Flask==3.0.3
mysql-connector-python==9.0.0